Fix the infinite loop in ThumbnailManager regarding /Network never getting removed. There were two problems:
a) /Network is technically a child of "/", but we want to force it to be a child of <EGPathRoot> instead.
b) Once that was done, there's still the problem that EGPathRoot objects aren't equal in arrays; EGPathRoot needed a hash function that returned a constant value for that to work.